Control Scheme Change for Matthew Teleop
This still needs to be tested. We need to find out at what point our motor has gone over the top of the robot and find out what the encoder reading at that point is. Once we have that we can plug it into the control switching part of the program.

Adds a fine mode to teleop. This makes it so that when you push a button the speed of the robot is decreased. This makes it much easier to grab things on the field and to make small precise movements.

Switches to tank mode